requirements.txt
crewai==0.114.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
gspread==6.1.2
python-dotenv==1.0.1
